+++ manage.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python
+
 import os
 import sys
 
 if __name__ == '__main__':
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'untitled.settings')
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'boilerplate.settings')
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
