+++ static/css/splash.css
@@ -1,5 +1,5 @@
 body {
-    background-image: url("../img/myy-wall.png");
+    background-image: url("../img/wall.png");
     background-repeat: repeat;
     background-size: cover;
 }
