+++ templates/base.html
@@ -15,7 +15,7 @@
 <body>
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark fixed-top bar">
     <div class="container">
-        <img class="navbar-brand" src="{% static "img/myy-logo.png" %}" alt="Yasin's Logo">
+        <img class="navbar-brand" src="{% static "img/logo.png" %}" alt="Yasin's Logo">
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Navigation">
             <span class="navbar-toggler-icon"></span>
         </button>
@@ -49,10 +49,10 @@
                 <h5 class="card-header">Hello World!</h5>
                 <div class="card-body">
                     <center>
-                        <img src="{% static "img/myy-draw.png" %}" alt="Yasin's Drawing">
+                        <img src="{% static "img/draw.png" %}" alt="Yasin's Drawing">
                     </center>
                     <hr>
-                    <a class="twitter-timeline" data-width=100% data-height="500" data-dnt="true" data-theme="dark" data-link-color="#FAB81E" href="https://twitter.com/myasiny?ref_src=twsrc%5Etfw">Tweets by myasiny13</a>
+                    <a class="twitter-timeline" data-width=100% data-height="500" data-dnt="true" data-link-color="#FAB81E"  href="https://twitter.com/myasiny?ref_src=twsrc%5Etfw">Tweets by myasiny</a> <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
                     <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
                 </div>
             </div>
@@ -62,7 +62,7 @@
 
 <footer class="py-5 bg-dark">
     <div class="container">
-        <p class="m-0 text-center text-white">&copy; M Yasin Y</p>
+        <p class="m-0 text-center text-white">2020 &copy; M Yasin Y</p>
     </div>
 </footer>
 </body>
