+++ templates/detail.html
@@ -3,7 +3,7 @@
 
 {% block content %}
     <script async src="https://static.addtoany.com/menu/page.js"></script>
-    <div class="a2a_kit a2a_kit_size_32 a2a_default_style paylas">
+    <div class="a2a_kit a2a_kit_size_32 a2a_default_style share-with">
         <a class="a2a_button_facebook"></a>
         <a class="a2a_button_twitter"></a>
         <a class="a2a_button_linkedin"></a>
