+++ templates/index.html
@@ -6,7 +6,7 @@
         <div class="card mb-4">
             <div class="card-body">
                 <h2 class="card-title">{{ post.title }}</h2>
-                <p class="card-text ozet">{{ post.body }}</p>
+                <p class="card-text short-description">{{ post.body }}</p>
                 <a href="{{ post.id }}" class="btn btn-primary">Read More &rarr;</a>
             </div>
             <div class="card-footer text-muted">
